1. add CMAKE_CXX_FLAGS 2. fix some warning 3. fix build error

+++ src/event/watcher/IWatcher.h
@@ -8,7 +8,7 @@
 
 #include "ev.h"
 
-using Event = unsigned long;
+using Event = int;
 static const Event EVENT_NONE = 0;
 
 class IWatcher {
